requirements.txt
aiofiles==23.1.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.0
altair==4.2.0
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
astor==0.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
auto-py-to-exe==2.33.0
Babel==2.10.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.73
blinker==1.5
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
colorlog==6.6.0
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.34
datasets==2.3.2
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.4
easydict==1.9
Eel==0.16.0
entrypoints @ file:///C:/ci/entrypoints_1649926621247/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.7.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.10.0
interpretdl==0.7.0
ipykernel @ file:///C:/ci/ipykernel_1647000978151/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651601994376/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315425835/work
jieba==0.42.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter-client @ file:///C:/ci/jupyter_client_1650534470695/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656285178/work
kiwisolver==1.4.2
librosa==0.8.1
linkify-it-py==2.0.0
llvmlite==0.36.0
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
mlxtend==0.21.0
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
natsort==8.1.0
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929372/work
networkx==2.8
numba==0.53.1
numpy==1.19.3
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.72
orjson==3.8.10
packaging==21.3
paddle2onnx==0.9.8
paddlefsl==1.1.0
paddlehub==2.1.0
paddlenlp==2.3.4
paddlepaddle==2.2.2
-e git+https://github.com/PaddlePaddle/PaddleSeg.git@2319feb42cd989a42a9ec154c6723deac6e44bb6#egg=paddleseg
paddleslim==1.1.1
paddlex==1.3.10
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pefile==2023.2.7
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.1
pkgutil_resolve_name==1.3.10
ply==3.11
pooch==1.6.0
ppgan==2.1.0
ppim==1.1.0
prettytable==3.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.0
psutil==5.9.1
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==8.0.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.7
pydeck==0.8.0b4
pydicom==2.3.1
pydub==0.25.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.15.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.18.1
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.2.0
QtPy==2.3.1
qudida==0.0.4
rarfile==4.0
requests==2.27.1
resampy==0.2.2
responses==0.18.0
rich==12.6.0
scikit-image==0.19.2
scikit-learn==0.24.0
scipy==1.8.1
seaborn==0.12.2
semantic-version==2.10.0
semver==2.13.0
sentencepiece==0.1.96
seqeval==1.2.2
Shapely==1.8.2
shiboken6==6.4.2
sip==6.7.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
SoundFile==0.10.3.post1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.26.1
streamlit==1.13.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
uc-micro-py==1.0.1
urllib3==1.26.9
uvicorn==0.21.1
validators==0.20.0
visualdl==2.3.0
watchdog==2.1.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==11.0.1
Werkzeug==2.1.2
wget==3.2
whichcraft==0.6.1
wincertstore==0.2
xlwt==1.3.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
zope.event==4.6
zope.interface==6.0
